+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * nostr: add `Kind::PeerToPeerOrder` variant ([Francisco Calderón])
 * nostr: add `CowTag` ([Yuki Kishimoto])
 * nostr: add `EventBorrow` ([Yuki Kishimoto])
+* database: add `Events::first_owned` and `Events::last_owned` ([Yuki Kishimoto])
 * pool: add `Relay::try_connect` ([Yuki Kishimoto])
 * pool: add `Relay::wait_for_connection` ([Yuki Kishimoto])
 * pool: add `RelayPool::try_connect` ([Yuki Kishimoto])